// src/app/api/model/externalTransferData.ts
/**
 * Licensed to the Apache Software Foundation (ASF) under one
 * or more contributor license agreements. See the NOTICE file
 * distributed with this work for additional information
 * regarding copyright ownership. The ASF licenses this file
 * to you under the Apache License, Version 2.0 (the
 * "License"); you may not use this file except in compliance
 * with the License. You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied. See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */

/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { ExternalTransferDataDetails } from './externalTransferDataDetails';
import { ExternalTransferLoanData } from './externalTransferLoanData';
import { ExternalTransferOwnerData } from './externalTransferOwnerData';


export interface ExternalTransferData { 
    details?: ExternalTransferDataDetails;
    effectiveFrom?: string;
    effectiveTo?: string;
    loan?: ExternalTransferLoanData;
    owner?: ExternalTransferOwnerData;
    previousOwner?: ExternalTransferOwnerData;
    purchasePriceRatio?: string;
    settlementDate?: string;
    status?: ExternalTransferData.StatusEnum;
    subStatus?: ExternalTransferData.SubStatusEnum;
    transferExternalGroupId?: string;
    transferExternalId?: string;
    transferId?: number;
}
export namespace ExternalTransferData {
    export const StatusEnum = {
        Active: 'ACTIVE',
        ActiveIntermediate: 'ACTIVE_INTERMEDIATE',
        Declined: 'DECLINED',
        Pending: 'PENDING',
        PendingIntermediate: 'PENDING_INTERMEDIATE',
        Buyback: 'BUYBACK',
        BuybackIntermediate: 'BUYBACK_INTERMEDIATE',
        Cancelled: 'CANCELLED'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
    export const SubStatusEnum = {
        BalanceZero: 'BALANCE_ZERO',
        BalanceNegative: 'BALANCE_NEGATIVE',
        SamedayTransfers: 'SAMEDAY_TRANSFERS',
        UserRequested: 'USER_REQUESTED',
        Unsold: 'UNSOLD'
    } as const;
    export type SubStatusEnum = typeof SubStatusEnum[keyof typeof SubStatusEnum];
}